Readback works ig

+++ Runtime/Systems/TerrainManagerSystem.cs
@@ -9,7 +9,7 @@ using Unity.Transforms;
 
 namespace jedjoud.VoxelTerrain {
     [UpdateInGroup(typeof(FixedStepTerrainSystemGroup))]
-    [UpdateBefore(typeof(TerrainOctreeJobSystem))]
+    [UpdateAfter(typeof(TerrainOctreeJobSystem))]
     public partial struct TerrainManagerSystem : ISystem {
         private NativeHashMap<OctreeNode, Entity> chunks;
         private Entity prototype;
@@ -23,7 +23,17 @@ namespace jedjoud.VoxelTerrain {
             EntityManager mgr = state.EntityManager;
             prototype = mgr.CreateEntity();
             mgr.AddComponent<LocalToWorld>(prototype);
-            mgr.AddComponent<TerrainChunkTag>(prototype);
+            mgr.AddComponent<TerrainChunk>(prototype);
+
+            mgr.AddComponent<TerrainChunkVoxels>(prototype);
+            mgr.AddComponent<TerrainChunkRequestReadbackTag>(prototype);
+            mgr.AddComponent<TerrainChunkRequestMeshingTag>(prototype);
+            mgr.AddComponent<TerrainChunkRequestCollisionTag>(prototype);
+
+            mgr.SetComponentEnabled<TerrainChunkVoxels>(prototype, false);
+            mgr.SetComponentEnabled<TerrainChunkRequestReadbackTag>(prototype, false);
+            mgr.SetComponentEnabled<TerrainChunkRequestMeshingTag>(prototype, false);
+            mgr.SetComponentEnabled<TerrainChunkRequestCollisionTag>(prototype, false);
         }
 
         [BurstCompile]
@@ -32,9 +42,13 @@ namespace jedjoud.VoxelTerrain {
             RefRW<TerrainOctree> _octree = SystemAPI.GetSingletonRW<TerrainOctree>();
             ref TerrainOctree octree = ref _octree.ValueRW;
 
-            if (!octree.pending && octree.handle.IsCompleted) {
+            if (!octree.pending && octree.handle.IsCompleted && octree.readyToSpawn) {
                 foreach (var node in octree.removed) {
                     if (chunks.TryGetValue(node, out var entity)) {
+                        TerrainChunkVoxels voxels = state.EntityManager.GetComponentData<TerrainChunkVoxels>(entity);
+                        voxels.inner.Dispose();
+
+
                         state.EntityManager.DestroyEntity(entity);
                         chunks.Remove(node);
                     }
@@ -46,17 +60,28 @@ namespace jedjoud.VoxelTerrain {
 
                     Entity chunk = state.EntityManager.Instantiate(prototype);
                     chunks.Add(node, chunk);
-                }
+                    state.EntityManager.SetComponentEnabled<TerrainChunkVoxels>(chunk, true);
+                    state.EntityManager.SetComponentEnabled<TerrainChunkRequestReadbackTag>(chunk, true);
 
-                // UnityEngine.Debug.Log($"A:{octree.added.Length} R:{octree.removed.Length}");
-                
+                    state.EntityManager.SetComponentData<TerrainChunkVoxels>(chunk, new TerrainChunkVoxels {
+                        inner = new NativeArray<Voxel>(VoxelUtils.VOLUME, Allocator.Persistent, NativeArrayOptions.UninitializedMemory)
+                    });
+                }
+                                
                 octree.continuous = true;
+                octree.readyToSpawn = false;
             }
         }
 
         [BurstCompile]
         public void OnDestroy(ref SystemState state) {
             chunks.Dispose();
+
+            foreach (var (voxels, entity) in SystemAPI.Query<TerrainChunkVoxels>().WithEntityAccess()) {
+                if (voxels.inner.IsCreated) {
+                    voxels.inner.Dispose();
+                }
+            }
         }
     }
 }
